+++ app/profile/ProfileTags.js
@@ -3,7 +3,7 @@
 /* globals promptError: false */
 import { useEffect, useState } from 'react'
 import api from '../../lib/api-client'
-import CheckableTag from '../../components/CheckableTag'
+import ProfileTag from '../../components/ProfileTag'
 import { prettyInvitationId, prettyId } from '../../lib/utils'
 
 export default function ProfileTags({ profileId, isSuperUser }) {
@@ -45,12 +45,7 @@ export default function ProfileTags({ profileId, isSuperUser }) {
   return (
     <div className={`tags-container ${tags.length ? 'mb-2' : ''}`}>
       {tags.map((tag, index) => (
-        <CheckableTag
-          key={index}
-          label={`${prettyInvitationId(tag.invitation)}${tag.label !== undefined ? ` "${tag.label}"` : ''}${tag.weight !== undefined ? ` (${tag.weight})` : ''}${!tag.invitation.startsWith(tag.signature) ? '' : ` by ${prettyId(tag.signature)}`}`}
-          checked={true}
-          onDelete={() => deleteTag(tag)}
-        />
+        <ProfileTag key={index} tag={tag} onDelete={() => deleteTag(tag)} />
       ))}
     </div>
   )
